Split transaction input gathering into assembler and parameter loader
Pure move: PositionAssembler owns snapshot assembly with unreported-
position adjustment behind one assemble entry point, and
TransactionParameterLoader owns model weights, limits, venues, and
instrument metadata behind one load entry point, while the service
keeps orchestration and the cash, fee, and liquidity lookups. A dead
repository field from a pre-split era goes with it, and the sixteen-
mock test decomposes into focused per-class tests wiring real
collaborators.

God classes 5 -> 4, PMD violations 135 -> 133, disconnected classes
453 -> 452.
