+++ metrics/scorecard.json
@@ -1,21 +1,21 @@
 {
-  "branchCoverage": 87.66,
+  "branchCoverage": 87.67,
   "classCouplingViolations": 34,
   "clockViolations": 0,
   "cognitiveComplexityViolations": 20,
   "compilerWarnings": 0,
   "deepNestingViolations": 0,
   "disabledTests": 5,
-  "disconnectedClasses": 453,
+  "disconnectedClasses": 452,
   "fqnViolations": 0,
-  "godClasses": 5,
+  "godClasses": 4,
   "interactionVerifyingTestClasses": 362,
-  "interactionVerifyingTestShare": 36.2,
+  "interactionVerifyingTestShare": 36.1,
   "lineCoverage": 84.04,
-  "longClasses": 2,
+  "longClasses": 1,
   "meanInstability": 0.473,
   "mocklessTestClasses": 462,
-  "mocklessTestShare": 46.2,
+  "mocklessTestShare": 46.1,
   "moduleCycles": 0,
   "moduleSliceTests": 0,
   "modulithViolations": 0,
@@ -26,23 +26,23 @@
   "overMockedTestClasses": 111,
   "overMockedTestShare": 11.1,
   "pitestSlicedModules": 52,
-  "pmdViolations": 135,
+  "pmdViolations": 133,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
     "CouplingBetweenObjects": 34,
-    "CyclomaticComplexity": 72,
-    "GodClass": 5,
+    "CyclomaticComplexity": 71,
+    "GodClass": 4,
     "TooManyMethods": 4
   },
-  "slowest10TestClassesSeconds": 115.9,
+  "slowest10TestClassesSeconds": 114.9,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 205,
+  "suiteWallTimeSeconds": 164,
   "testClassesOver10Seconds": 4,
-  "testClassesTotal": 1001,
+  "testClassesTotal": 1003,
   "testIntegrationShare": 6.7,
   "testSliceShare": 13.8,
-  "testUnitShare": 79.5,
-  "top15ClassLines": 6649,
+  "testUnitShare": 79.6,
+  "top15ClassLines": 6478,
   "totalPackages": 227,
   "unmarkedPackages": 0
 }
